Update README.md
A Detective Story – “The Last Job”

This is a story about a detective sent out to investigate an old crime scene in an old haunted 
house on the hill. He arrives and faces a series of spooky challenges in the different rooms 
of the house. Solving or not solving a puzzle could mean life or death for him. 
The whole experience of this investigation changes the detective forever. So will this be his 
‘last job’ because he has found lots of treasure or because he is not alive to do another job?

To run the game type py LASTJOBFinal.py 

Project Management Review 

Overall I felt the project went well and I learnt a lot about collaborative working as well as coding in Python. 
We worked together as a team of three and shared our ideas for the game quickly and effectively. We 
talked about the design and levels of the game and allocated individual games rooms to different members 
for coding. Then we tested our game modules and user tested the game, made improvements, and 
created a new version that was again tested and any bugs and glitches were fed back to the team and 
fixed.

Stand-up’s are very helpful, focussed and powerful. Of course, on zoom they are ‘sit-downs’ but the focus 
is still the same. They helped us summarise our individual and collective progress with the project, share 
ideas, allocate tasks, and invite feedback and questions. We also used Slack and Trello to communicate 
when not on a zoom call. 

Kanban Planning was great for managing the workflow of the project, task allocation and updates. It acts 
like a glorified to-do list for a project broken down into individual tasks. There were allocated to a team 
member. They then did the task and updated the Kanban board on Trello. This was easy to manage and 
feedback to the team. 

Efficiency of the Team as a whole

The team worked well as a whole, we all communicated well, and each member played to their strengths. We 
had very few issues with time management and any issues were resolved quickly by the team.

How management of the project could be improved
I watched a youtube video on ‘Kanban’ and thought it could be used for breaking down individual tasks before 
doing the tasks and charting the progress.

Some of the work was done out of hours by members of the team. There were no scheduled meetings out of 
hours. Instead we used Slack to share our code and comments. An out-of-hours zoom meeting on a couple of 
occasions would have been a good idea.

Frontend v. Backend

Frontend development with HTML and CSS would allow us to market and promote our game 
online to the world. It would also allow us to add graphics, music, and visual interaction to 
enrichen the game. The player would have a more immersive experience and it would be 
available to play 24/7 by a massive global audience. 

Dedicating more time to the front-end development means learning and applying HTML, 
CSS, Javascript and other tools. Each has a learning curve and the more technologies that 
are used the more complex development and testing becomes. There would be less time to 
develop the game in Python meaning less gameplay, less time for testing, and potentially 
more bugs.

My Review of ‘Last Job’

‘Last Job’ is the last case the detective will solve. There is an art and a science to creating a 
game using just the Python terminal, no graphics, and no sound. Instead we used ASCII art to 
visualise the scenes and good old-fashioned storytelling to illustrate the adventure with words 
and puzzles for the player to solve. 

I feel the project has gone well overall and we are ahead on testing. I managed the project, 
leading daily stand-up meetings, the Trello board, and I wrote the game code for the ghost 
room and I improved the code for the skeleton room. We communicated regularly via zoom, 
Slack and Trello. 

Most of the code was done and tested early. I also added extra features to the code for the 
skeleton and ghost rooms to make them more playable and challenging. I tested my code and 
the whole game code and found coding in Python enjoyable and immersive.

Game Testing - How Did It Go?

How our game was tested?

The game was tested when a version was ready and uploaded to Slack. Each option or 
outcome was tested for errors and the game was played from the users perspective. Testing 
was re-done each time there were code changes. 

Was testing effective?

Testing was effective. It highlighted issues with the game play and some team members 
found errors in the code that were not found by the person writing the code. Testing makes 
the game an enjoyable experience to play without bugs.

The Team

Tim Stokes
Project manager, coder, and tester

David Lynch
Senior coder, tester, and wireframe creator

Marian Krchlik
Coder, tester, and ASCII art designer